+++ apps/web/src/features/dashboard/pages/page1.tsx
@@ -1,8 +1,11 @@
+import { useDashboardPage2Route } from '../../../routes/routes.protocol';
 import { useCounterStore } from '../store/counter-store.hook';
 
 export function DashboardPage1() {
     const counterStore = useCounterStore();
     const count = counterStore.use.count();
+
+    const page2Route = useDashboardPage2Route();
 
     const handleIncrease = () => {};
 
@@ -12,6 +15,10 @@ export function DashboardPage1() {
 
     const handleRedo = () => {};
 
+    const goToP2 = () => {
+        page2Route.navigate({});
+    };
+
     return (
         <div>
             <div>dashboard page1</div>
@@ -20,7 +27,7 @@ export function DashboardPage1() {
             <button onClick={handleDecrease}>decrease</button>
             <button onClick={handleUndo}>undo</button>
             <button onClick={handleRedo}>redo</button>
-            <button> go</button>
+            <button onClick={goToP2}> go</button>
         </div>
     );
 }