+++ packages/viness-react/src/router/route.ts
@@ -10,15 +10,11 @@ import { NavigateOptions, PathParam } from './types';
 import { IVinessRoute } from './route.protocol';
 import type { VinessRouteMetadata } from './route.protocol';
 import type { IVinessRouter } from './router.protocol';
-import { Inject } from '@viness/core';
-import { VinessRouter } from './router';
 
 /**
  * Route
  */
 export class VinessRoute<Path extends string> implements IVinessRoute<Path> {
-    private static VROUTE = true;
-
     id?: string;
     path: Path;
     element: ReactNode;
@@ -28,25 +24,12 @@ export class VinessRoute<Path extends string> implements IVinessRoute<Path> {
     hasErrorBoundary?: boolean;
     caseSensitive?: boolean;
     index?: boolean | undefined;
-    lazy?: LazyRouteFunction<IndexRouteObject> | LazyRouteFunction<NonIndexRouteObject> | undefined;
+    lazy?: LazyRouteFunction<IndexRouteObject> | LazyRouteFunction<NonIndexRouteObject>;
     loader?: LoaderFunction | undefined;
-    children?: VinessRouteMetadata<Path>[] | undefined;
 
-    constructor(metadata: VinessRouteMetadata<Path>, @Inject(VinessRouter) private router: IVinessRouter) {
-        const {
-            id,
-            path,
-            index,
-            element,
-            errorElement,
-            Component,
-            ErrorBoundary,
-            hasErrorBoundary,
-            caseSensitive,
-            children,
-            lazy,
-            loader
-        } = metadata;
+    constructor(metadata: VinessRouteMetadata<Path>, public router: IVinessRouter) {
+        const { id, path, index, element, errorElement, Component, ErrorBoundary, hasErrorBoundary, caseSensitive, lazy, loader } =
+            metadata;
 
         this.id = id;
         this.path = path;
@@ -59,11 +42,10 @@ export class VinessRoute<Path extends string> implements IVinessRoute<Path> {
         this.caseSensitive = caseSensitive;
         this.lazy = lazy;
         this.loader = loader;
-        this.children = children;
     }
 
     navigate(params: { [key in PathParam<Path>]: string | null }, option?: NavigateOptions) {
-        const toPath = this.generatePath(params);
+        const toPath = this.generateFullPath(params);
         this.router.navigate(toPath, option);
     }
 
@@ -71,7 +53,7 @@ export class VinessRoute<Path extends string> implements IVinessRoute<Path> {
         return this.router.getParams<Path>();
     }
 
-    generatePath(params?: { [key in PathParam<Path>]: string | null }) {
+    generateFullPath(params?: { [key in PathParam<Path>]: string | null }) {
         const toPath = innerGeneratePath(this.path || '', params || ({} as any));
         return toPath;
     }